Extract common test tasks into new common helper object
