+++ app/tests/common/helper.js
@@ -1,0 +1,54 @@
+const config = require('config');
+const express = require('express');
+const log = require('npmlog');
+const Problem = require('api-problem');
+
+/** This class provides helper utilities that are commonly used in tests */
+const helper = {
+  /**
+   * Creates a stripped-down simple Express server object
+   * @param {string} basePath The path to mount the `router` on
+   * @param {object} router An express router object to mount
+   * @returns A simple express server object with `router` mounted to `basePath`
+   */
+  expressHelper: (basePath, router) => {
+    const app = express();
+
+    app.use(express.json());
+    app.use(express.urlencoded({
+      extended: false
+    }));
+    app.use(basePath, router);
+
+    // Handle 500
+    // eslint-disable-next-line no-unused-vars
+    app.use((err, _req, res, _next) => {
+      if (err instanceof Problem) {
+        err.send(res);
+      } else {
+        new Problem(500, {
+          details: (err.message) ? err.message : err
+        }).send(res);
+      }
+    });
+
+    // Handle 404
+    app.use((_req, res) => {
+      new Problem(404).send(res);
+    });
+
+    return app;
+  },
+
+  /**
+   * Configures an npmlog instance to have debug level logging and the right log level
+   */
+  logHelper: () => {
+    log.level = config.get('server.logLevel');
+    log.addLevel('debug', 1500, {
+      fg: 'cyan'
+    });
+  }
+};
+
+module.exports = helper;
