+++ app/tests/unit/routes/v1.spec.js
@@ -1,18 +1,13 @@
-const express = require('express');
 const request = require('supertest');
 
+const helper = require('../../common/helper');
 const router = require('../../../src/routes/v1');
 
 jest.mock('bull');
 
 // Simple Express Server
 const basePath = '/api/v1';
-const app = express();
-app.use(express.json());
-app.use(express.urlencoded({
-  extended: false
-}));
-app.use(basePath, router);
+const app = helper.expressHelper(basePath, router);
 
 describe('GET /api/v1', () => {
   it('should return all available endpoints', async () => {
