+++ app/tests/unit/routes/v1/status.spec.js
@@ -1,7 +1,6 @@
-const express = require('express');
 const request = require('supertest');
-const Problem = require('api-problem');
 
+const helper = require('../../../common/helper');
 const router = require('../../../../src/routes/v1/status');
 const statusComponent = require('../../../../src/components/status');
 
@@ -9,24 +8,7 @@ jest.mock('bull');
 
 // Simple Express Server
 const basePath = '/api/v1/status';
-const app = express();
-app.use(express.json());
-app.use(express.urlencoded({
-  extended: false
-}));
-app.use(basePath, router);
-
-// Handle 500
-// eslint-disable-next-line no-unused-vars
-app.use((err, _req, res, _next) => {
-  if (err instanceof Problem) {
-    err.send(res);
-  } else {
-    new Problem(500, {
-      details: (err.message) ? err.message : err
-    }).send(res);
-  }
-});
+const app = helper.expressHelper(basePath, router);
 
 const id = 'id';
